crud posts finalizado

+++ open/src/main/java/br/com/open/api/open/DTO/CommentsDTO.java
@@ -1,4 +1,4 @@
 package br.com.open.api.open.DTO;
 
-public record CommentsDTO(int postId, int id, String name, String email, String body) {
+public record CommentsDTO(long postId, long id, String name, String email, String body) {
 }
