+++ open/src/main/java/br/com/open/api/open/DTO/PostDTO.java
@@ -1,4 +1,5 @@
 package br.com.open.api.open.DTO;
 
-public record PostDTO(int userId, long id, String title, String body) {
+
+public record PostDTO(long userId, long id, String title, String body) {
 }
